+++ packages/frameworks/vue/src/date-picker/use-date-picker.ts
@@ -7,7 +7,10 @@ import { useId } from '../utils'
 
 export interface UseDatePickerProps
   extends Optional<
-    Omit<datePicker.Context, 'value' | 'min' | 'max' | 'focusedValue' | 'open.controlled'>,
+    Omit<
+      datePicker.Context,
+      'value' | 'min' | 'max' | 'parse' | 'focusedValue' | 'open.controlled'
+    >,
     'id'
   > {
   /**